pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "antspymm"
version = "1.5.0"
description = "multi-channel/time-series medical image processing with antspyx"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Apache-2.0" }
authors = [
    { name = "Avants, Gosselin, Tustison, Reardon", email = "stnava@gmail.com" }
]
classifiers = [
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent"
]
dependencies = [
    "h5py>=2.10.0",
    "numpy>=1.19.4",
    "pandas>=1.0.1",
    "antspyx>=0.4.2",
    "antspynet>=0.2.8",
    "antspyt1w>=0.9.3",
    "pathlib",
    "dipy",
    "nibabel",
    "scipy",
    "siq",
    "scikit-learn"
]

[tool.setuptools.packages]
find = { include = ["antspymm"] }

[tool.setuptools]
license-files = ["LICENSE"]